organize components folder, update link on Navbar
