--- src/App.jsx
@@ -1,18 +1,0 @@
-// import Navbar from "../components/Navbar";
-// import Footer from "../components/DonateNow";
-// import { Outlet } from "react-router-dom";
-// function App() {
-  
-//   return (
-//     <>
-//       <div > 
-//             <Navbar />      
-//             <Outlet/>
-//             <Footer/>
-      
-//         </div>
-//     </>
-//   )
-// }
-
-// export default App
