+++ src/components/About/Founder.jsx
@@ -1,0 +1,7 @@
+
+
+export default function Founder() {
+  return (
+    <div>Founder</div>
+  )
+}
