+++ src/components/DomesticViolence/SolutionsOfDV.jsx
@@ -1,4 +1,4 @@
-import {s1,s2,s3,s4,s5,s6,s7,s8,s9,s10,s11,s12} from '../assets/';
+import {s1,s2,s3,s4,s5,s6,s7,s8,s9,s10,s11,s12} from '../../assets';
 
 export default function SolutionsOfDV() {
   return (
